pyproject.toml
[tool.poetry]
name = "issue-system-backend"
version = "0.1.0"
description = "Issue Management System Backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1"
djangorestframework = "^3.15"
mysqlclient = "^2.2"
pendulum = "^3.0"
drf-spectacular = "^0.27"
django-cors-headers = "^4.3"
celery = "^5.4"
redis = "^5.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
ruff = "^0.1"
black = "^24.0"
ipython = "^8.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

